// src/app/shared/model/user.ts
export class User {
    $key: string;
  userName: string;
  emailId: string;
  password: string;
  location: {
    lat: number;
    lon: number;
    };
    phoneNumber: string;
    createdOn: string;
    isAdmin: boolean;
    photo: string;
  
}
export class UserDetails{
    $key: string;
  firstName: string;
  lastName: string;
  userName: string;
  emailId: string;
  address1: string;
  address2: string;
  country: string;
  state: string;
  zip: number;
}